Refactor: Reorganize the code in modmesh/view
* Add RStaticMesh.cpp for the non template code.
* Add modmesh/view/view.hpp as the module interface header file.
