+++ cpp/modmesh/view/CMakeLists.txt
@@ -5,16 +5,20 @@ cmake_minimum_required(VERSION 3.16)
 
 set(MODMESH_VIEW_HEADERS
     ${CMAKE_CURRENT_SOURCE_DIR}/R3DWidget.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/RApplication.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/RMainWindow.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/RPythonText.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/RStaticMesh.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/view.hpp
     CACHE FILEPATH "" FORCE
 )
 
 set(MODMESH_VIEW_SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/R3DWidget.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/RApplication.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/RMainWindow.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/RPythonText.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/RStaticMesh.cpp
     CACHE FILEPATH "" FORCE
 )
