+++ cpp/modmesh/view/RStaticMesh.hpp
@@ -73,78 +73,6 @@ private:
 
 }; /* end class RStaticMesh */
 
-inline RStaticMesh::RStaticMesh(std::shared_ptr<StaticMesh> const & static_mesh, Qt3DCore::QNode * parent)
-    : Qt3DCore::QEntity(parent)
-    , m_geometry(new Qt3DCore::QGeometry(this))
-    , m_renderer(new Qt3DRender::QGeometryRenderer())
-    , m_material(new Qt3DExtras::QDiffuseSpecularMaterial())
-{
-    update_geometry(*static_mesh);
-    m_renderer->setGeometry(m_geometry);
-    m_renderer->setPrimitiveType(Qt3DRender::QGeometryRenderer::Lines);
-    addComponent(m_renderer);
-    addComponent(m_material);
-}
-
-inline void RStaticMesh::update_geometry_impl(StaticMesh const & mh, Qt3DCore::QGeometry * geom)
-{
-    auto * buf = new Qt3DCore::QBuffer(geom);
-    {
-        QByteArray barray;
-        barray.resize(3 * mh.nnode() * sizeof(float));
-        float * ptr = reinterpret_cast<float *>(barray.data());
-        for (uint32_t ind = 0; ind < mh.nnode(); ++ind)
-        {
-            *ptr++ = mh.ndcrd(ind, 0);
-            *ptr++ = mh.ndcrd(ind, 1);
-            *ptr++ = (3 == mh.ndim()) ? mh.ndcrd(ind, 2) : 0;
-        }
-        buf->setData(barray);
-    }
-
-    auto * vertices = new Qt3DCore::QAttribute(geom);
-    vertices->setName(Qt3DCore::QAttribute::defaultPositionAttributeName());
-    vertices->setAttributeType(Qt3DCore::QAttribute::VertexAttribute);
-    vertices->setVertexBaseType(Qt3DCore::QAttribute::Float);
-    vertices->setVertexSize(3);
-    vertices->setBuffer(buf);
-    vertices->setByteStride(3 * sizeof(float));
-    vertices->setCount(mh.nnode());
-
-    geom->addAttribute(vertices);
-
-    // FIXME: This is naive implementation and creates a lot of duplicated
-    // edges.  The StaticMesh should provide a set of unique edges.
-    uint32_t nend = 0;
-    for (uint32_t ifc = 0; ifc < mh.nface(); ++ifc)
-    {
-        nend += mh.fcnds(ifc, 0);
-    }
-
-    buf = new Qt3DCore::QBuffer(geom);
-    {
-        QByteArray barray;
-        barray.resize(nend * sizeof(uint32_t));
-        auto * indices = reinterpret_cast<uint32_t *>(barray.data());
-        for (uint32_t ifc = 0; ifc < mh.nface(); ++ifc)
-        {
-            for (int32_t inf = 1; inf <= mh.fcnds(ifc, 0); ++inf)
-            {
-                *indices++ = mh.fcnds(ifc, inf);
-            }
-        }
-        buf->setData(barray);
-    }
-
-    auto * indices = new Qt3DCore::QAttribute(geom);
-    indices->setVertexBaseType(Qt3DCore::QAttribute::UnsignedInt);
-    indices->setAttributeType(Qt3DCore::QAttribute::IndexAttribute);
-    indices->setBuffer(buf);
-    indices->setCount(nend);
-
-    geom->addAttribute(indices);
-}
-
 } /* end namespace modmesh */
 
 // vim: set ff=unix fenc=utf8 et sw=4 ts=4 sts=4:
